chore: lazy-image migrated to ts
